requirements.txt
pandas==2.2.3
phidata==2.7.5
streamlit==1.41.1
